+++ dzVents/runtime/device-adapters/rgbw_device.lua
@@ -1,5 +1,3 @@
-local TimedCommand = require('TimedCommand')
-
 return {
 
 	baseType = 'device',